fixed async error

+++ post.js
@@ -124,7 +124,7 @@ module.exports = {
                 let serverInfo;
                 allServers.map(val => val.serverID === server.id ? serverInfo = val : undefined)
 
-                server.channels.cache.map(channel => {
+                server.channels.cache.map(async(channel) => {
                     if (channel.id === serverInfo.sbGuideChannelID) {
                         // channel.messages.fetch({around: guideMessage[0][serverInfo.serverID], limit: 1})
 					    //     .then(msg => {
@@ -135,9 +135,8 @@ module.exports = {
 
                         channel.messages.fetch({around: guideMessage[0].sbTable, limit: 1})
                             .then(msg => {
-                                if (guideMessage[0].sbTable != msg.id) break;
+                                if (guideMessage[0].sbTable == msg.id) msg.first().delete()
                                 //if the msg id fetched doesn't match, assume the message is lost/deleted
-                                msg.first().delete()
                             })
 
                         await globalFunctions.tableOfContents("Skyblock", serverID)
